requirements.txt
celery==5.1.2
Django==3.1.4
django-cors-headers==3.4.0
django-filter==2.3.0
django-imagekit==4.0
djangorestframework==3.12.2
drf-autodocs==0.4.3
openpyxl==3.0.4
pilkit==2.0
Pillow==8.3.2
